# docs/developer-docs/public/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tari_indexer",
    "description": "The Tari indexer implementation",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.37.0"
  },
  "paths": {
...
    "/transaction-receipts": {
      "get": {
        "tags": [
          "handlers::transaction_receipts"
        ],
        "description": "List transaction receipts",
        "operationId": "list_transaction_receipts",
        "responses": {
          "200": {
            "description": "List of transaction receipts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTransactionReceiptsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list transaction receipts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transaction-receipts/{address}": {
      "get": {
        "tags": [
          "handlers::transaction_receipts"
        ],
        "description": "Get the transaction receipt by its address",
        "operationId": "get_transaction_receipt",
        "responses": {
          "200": {
            "description": "Transaction receipt found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionReceiptResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transaction receipt not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch transaction receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/transactions/events/stream": {
      "get": {
        "tags": [
          "handlers::transaction_events"
        ],
        "description": "SSE stream of template-emitted transaction events. Supports catch-up via the `after_id` query parameter or `Last-Event-ID` header.",
        "operationId": "sse_transaction_events",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "description": "Filter by event topic",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "substate_id",
            "in": "query",
            "description": "Filter by substate ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "template_address",
            "in": "query",
            "description": "Filter by template address",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_address",
            "in": "query",
            "description": "Filter by resource address (derived from substate_id for std.resource.* events, or from the `resource_address` payload entry for std.vault.deposit / std.vault.withdraw)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after_id",
            "in": "query",
            "description": "Resume from this event ID (exclusive)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {}
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetTransactionReceiptResponse": {
        "type": "object",
        "required": [
          "receipt"
        ],
        "properties": {
          "receipt": {
            "type": "object"
          }
        }
      },
...
      "ListTransactionReceiptsResponse": {
        "type": "object",
        "required": [
          "receipts"
        ],
        "properties": {
          "receipts": {
            "type": "array",
            "items": {
              "type": "array",
              "items": false,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "object"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}